Refactor to a single main purpose for each group
