+++ cmd/fyne_demo/tutorials/canvas.go
@@ -17,8 +17,8 @@ func rgbGradient(x, y, w, h int) color.Color {
 	return color.NRGBA{uint8(255 - b), uint8(g), uint8(b), 0xff}
 }
 
-// GraphicsScreen loads a graphics example panel for the demo app
-func GraphicsScreen(_ fyne.Window) fyne.CanvasObject {
+// canvasScreen loads a graphics example panel for the demo app
+func canvasScreen(_ fyne.Window) fyne.CanvasObject {
 	gradient := canvas.NewHorizontalGradient(color.NRGBA{0x80, 0, 0, 0xff}, color.NRGBA{0, 0x80, 0, 0xff})
 	go func() {
 		for {
@@ -31,7 +31,8 @@ func GraphicsScreen(_ fyne.Window) fyne.CanvasObject {
 			canvas.Refresh(gradient)
 		}
 	}()
-	content := fyne.NewContainerWithLayout(layout.NewGridWrapLayout(fyne.NewSize(90, 90)),
+
+	return fyne.NewContainerWithLayout(layout.NewGridWrapLayout(fyne.NewSize(90, 90)),
 		canvas.NewImageFromResource(theme.FyneLogo()),
 		&canvas.Rectangle{FillColor: color.NRGBA{0x80, 0, 0, 0xff},
 			StrokeColor: color.NRGBA{0xff, 0xff, 0xff, 0xff},
@@ -45,6 +46,4 @@ func GraphicsScreen(_ fyne.Window) fyne.CanvasObject {
 		gradient,
 		canvas.NewRadialGradient(color.NRGBA{0x80, 0, 0, 0xff}, color.NRGBA{0, 0x80, 0x80, 0xff}),
 	)
-
-	return fyne.NewContainerWithLayout(layout.NewAdaptiveGridLayout(2), content, IconsPanel())
 }
