+++ cmd/fyne_demo/tutorials/data.go
@@ -17,30 +17,42 @@ var (
 			"Welcome to the Fyne toolkit tutorials.\nChoose an area from the options on the left.",
 			welcomeScreen,
 		},
-		"graphics": {"Graphics",
-			"See the canvas and graphics capabilities",
-			GraphicsScreen,
+		"canvas": {"Canvas",
+			"See the canvas capabilities",
+			canvasScreen,
 		},
-		"widgets": {"Widgets",
-			"Browse the toolkit widgets",
-			WidgetScreen,
+		"icons": {"Theme Icons",
+			"Browse the embedded icons",
+			iconScreen,
 		},
 		"containers": {"Containers",
 			"Containers and collections",
-			ContainerScreen,
+			containerScreen,
+		},
+		"widgets": {"Widgets",
+			"Browse the toolkit widgets",
+			widgetScreen,
+		},
+		"collections": {"Collection Widgets",
+			"Learn about the collection widgets",
+			collectionScreen,
+		},
+		"dialogs": {"Dialogs",
+			"Work with dialogs",
+			dialogScreen,
 		},
 		"windows": {"Windows",
-			"Window and dialog demos",
-			DialogScreen,
+			"Window function demo",
+			windowScreen,
 		},
 		"advanced": {"Advanced",
 			"Debug and advanced information",
-			AdvancedScreen,
+			advancedScreen,
 		},
 	}
 
 	// TutorialIndex  defines how our tutorials should be laid out in the index tree
 	TutorialIndex = map[string][]string{
-		"": {"welcome", "graphics", "widgets", "containers", "windows", "advanced"},
+		"": {"welcome", "canvas", "icons", "widgets", "collections", "containers", "dialogs", "windows", "advanced"},
 	}
 )
